Uncommented alert...hopefully this works

+++ assets/scripts/welcome.js
@@ -30,10 +30,10 @@ document.getElementById("runners").onclick = runnerClicked;
 function runnerClicked() {
   var bodyColor = window.getComputedStyle(document.body).getPropertyValue('background-color');
   var movingRunner = window.getComputedStyle(document.getElementById("one")).getPropertyValue("display");
-  // if (bodyColor == "#222" || movingRunner == "block") {
-  //   window.alert("Thank you for visiting!");
-  //   return;
-  // }
+  if (bodyColor == "#222" || movingRunner == "block") {
+    window.alert("Thank you for visiting!");
+    return;
+  }
   wipeOutWelcome();
   slideInPageTwo();
   sunReflection();
